fix: vgm2la: `-r` option can take a real number.
In a typical example, if the VGM was ripped on a PAL/SECAM machine (e.g.
European MSX), `-r 50` applies, and if ripped on an NTSC machine (e.g.
Japanese MSX), `-r 60` or `-r 59.94` applies.

+++ tools/vgm2la/src/AudioSample.hpp
@@ -19,7 +19,7 @@
 #include "VGMHeader.hpp"
 
 struct AudioSample {
-  int rate;
+  double rate;
   uint8_t soundchip_enable;
   uint8_t channel_enable;
